+++ src/Phormium/Config.php
@@ -43,9 +43,11 @@ class Config
         return self::$config['databases'][$name];
     }
 
-    public static function addDatabase($name, $dsn, $username = null, $password = null)
+    public static function addDatabase($name, $dsn, $username = null, $password = null, array $attributes = [])
     {
-        self::$config['databases'][$name] = compact('dsn', 'username', 'password');
+        $attributes = self::processAttributes($name, $attributes);
+
+        self::$config['databases'][$name] = compact('dsn', 'username', 'password', 'attributes');
     }
 
     /** Load configuration from array or file. */
@@ -68,7 +70,71 @@ class Config
 
         $processor = new Processor();
         $tree = self::getConfigTreeBuilder()->buildTree();
-        self::$config = $processor->process($tree, array($config));
+        $config = $processor->process($tree, array($config));
+        self::$config = self::processConfig($config);
+    }
+
+    /**
+     * Process and validate a configuration array.
+     */
+    private static function processConfig(array $config)
+    {
+        foreach ($config['databases'] as $name => $dbConfig) {
+            $config['databases'][$name] = self::processDbConfig($name, $dbConfig);
+        }
+
+        return $config;
+    }
+
+    private static function processDbConfig($name, $config)
+    {
+        $config['attributes'] = self::processAttributes($name, $config['attributes']);
+
+        return $config;
+    }
+
+    private static function processAttributes($dbName, $attributes)
+    {
+        $processed = [];
+
+        foreach ($attributes as $name => $value) {
+            try {
+                $procName = self::processConstant($name, false);
+            } catch (\Exception $ex) {
+                throw new \Exception("Invalid attribute \"$name\" specified in configuration for database \"$dbName\".");
+            }
+
+            try {
+                $procValue = self::processConstant($value, true);
+            } catch (\Exception $ex) {
+                throw new \Exception("Invalid value \"$value\" given for attribute \"$name\", in configuration for database \"$dbName\".");
+            }
+
+            $processed[$procName] = $procValue;
+        }
+
+        return $processed;
+    }
+
+    private static function processConstant($value, $allowScalar = false)
+    {
+        // If the value is an integer, assume it's a PDO::* constant value
+        // and leave it as-is
+        if (is_integer($value)) {
+            return $value;
+        }
+
+        // If it's a string which starts with "PDO::", try to find the
+        // corresponding PDO constant
+        if (is_string($value) && substr($value, 0, 5) === 'PDO::' && defined($value)) {
+            return constant($value);
+        }
+
+        if ($allowScalar && is_scalar($value)) {
+            return $value;
+        }
+
+        throw new \Exception("Invalid constant value");
     }
 
     private static function getConfigTreeBuilder()
@@ -88,6 +154,12 @@ class Config
                             ->end()
                             ->scalarNode('password')
                                 ->defaultNull()
+                            ->end()
+                            ->arrayNode('attributes')
+                                ->useAttributeAsKey('name')
+                                ->prototype('scalar')
+                                    ->isRequired()
+                                ->end()
                             ->end()
                         ->end()
                     ->end()